Renamed Header to HeaderBase for more clarity
